add solution to  (the first part of) 11
Add the "stupid" version of how to get the result.

Although the problem can be solved using this version by brute-forcing the first few n x n grids, it is quite ugly and pretty much the same as this one, so I'll add it here when I've got a better one.